I made "addNode" more robust in NameServerDatabase.java. Now nodes that happen to have the same Hashcode don't overide each other.
Added a way to delete nodes.

Made a lot of tests for these.

Also notices a small mistake in Mapping.java
When casting from a double to an int everything after the comma gets dropped (so 1.9 becomes 1)

This made it nearly impossible to get 32768 out of NameToHash.convert, 32767 was the highest I could get. fixed it by adding Math.round before casting to int (still needs to vast because it returns a long)