added all API and started with a test.
I have not testes all the API util methods and just want to continue with the shutdown. I think it's enough to keeep going on with each our parts. If you find any mistakes in these methodes then fix and merge with the main as soon as you can or something.